Fix implicit cast

+++ jive_layouts/layout/gui-items/jive_ContainerItemChild.cpp
@@ -201,7 +201,8 @@ namespace jive
 
         if (auto* container = getTopLevelDecorator().toType<ContainerItem>())
         {
-            if (getComponent()->getWidth() < idealWidth.get() || getComponent()->getHeight() < idealHeight.get())
+            if (static_cast<float>(getComponent()->getWidth()) < idealWidth.get()
+                || static_cast<float>(getComponent()->getHeight()) < idealHeight.get())
                 container->updateIdealSizeWithinConstraints();
         }
     }
